+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Unreleased
 ----------
+- Added example illustrating how to submit a limit order
 - Bumped `http-endpoint` dependency to `0.2`